Fixing codestyle issues, synchronizing currentOrder in SimpleTaxi
