+++ TaxiService/src/main/java/taxiservice/Order.java
@@ -1,4 +1,4 @@
-package taxi_service;
+package taxiservice;
 
 public interface Order {
     void processOrder();
